refactor(examples): use local modules
