+++ _examples/common/consumer/endpoints.go
