+++ _examples/common/consumer/request_response.go
