+++ _examples/common/consumer/service.go
