+++ _examples/common/consumer/storage.go
