+++ _examples/common/consumer/transport_http.go
