+++ _examples/common/domain/constants.go
