+++ _examples/common/domain/event.go
