+++ _examples/common/go.mod
@@ -1,0 +1,13 @@
+module github.com/alebabai/go-kit-kafka/examples/common
+
+go 1.16
+
+require (
+	github.com/alebabai/go-kit-kafka v0.0.0
+	github.com/go-kit/kit v0.13.0
+	github.com/go-kit/log v0.2.1
+	github.com/google/uuid v1.2.0
+	github.com/gorilla/mux v1.8.0
+)
+
+replace github.com/alebabai/go-kit-kafka => ../../
