+++ _examples/common/producer/endpoints.go
