+++ _examples/common/producer/generator.go
