+++ _examples/common/producer/middleware.go
