+++ _examples/common/producer/service.go
