+++ _examples/common/producer/transport_kafka.go
