+++ _examples/confluent/cmd/consumer/main.go
@@ -8,8 +8,8 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/Shopify/sarama"
 	kafkatransport "github.com/alebabai/go-kit-kafka/kafka/transport"
+	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/go-kit/kit/transport"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
@@ -17,7 +17,7 @@ import (
 	"github.com/alebabai/go-kit-kafka/examples/common/consumer"
 	"github.com/alebabai/go-kit-kafka/examples/common/domain"
 
-	"github.com/alebabai/go-kit-kafka/examples/sarama/pkg/consumer/kafka/adapter"
+	"github.com/alebabai/go-kit-kafka/examples/confluent/pkg/consumer/kafka/adapter"
 )
 
 func fatal(logger log.Logger, err error) {
@@ -43,6 +43,7 @@ func main() {
 		logger = level.NewInjector(logger, level.InfoValue())
 		logger = log.With(logger, "ts", log.DefaultTimestampUTC)
 	}
+
 	_ = logger.Log("msg", "initialization of the application")
 
 	_ = logger.Log("msg", "initialize services")
@@ -76,31 +77,23 @@ func main() {
 
 	var kafkaListener *adapter.Listener
 	{
-		cfg := sarama.NewConfig()
-		cfg.Consumer.Offsets.Initial = sarama.OffsetOldest
-		cfg.Consumer.Offsets.AutoCommit.Enable = true
-
 		brokerAddr := domain.BrokerAddr
 		if v, ok := os.LookupEnv("BROKER_ADDR"); ok {
 			brokerAddr = v
 		}
 
-		client, err := sarama.NewClient(
-			[]string{brokerAddr},
-			cfg,
-		)
+		c, err := kafka.NewConsumer(&kafka.ConfigMap{
+			"bootstrap.servers":  brokerAddr,
+			"group.id":           domain.GroupID,
+			"enable.auto.commit": true,
+		})
 		if err != nil {
-			fatal(logger, fmt.Errorf("failed to init kafka client: %w", err))
-		}
-
-		consumerGroup, err := sarama.NewConsumerGroupFromClient(domain.GroupID, client)
-		if err != nil {
-			fatal(logger, fmt.Errorf("failed to init kafka consumer group: %w", err))
+			fatal(logger, fmt.Errorf("failed to init kafka consumer: %w", err))
 		}
 
 		defer func() {
-			if err := consumerGroup.Close(); err != nil {
-				fatal(logger, fmt.Errorf("failed to close kafka consumer group: %w", err))
+			if err := c.Close(); err != nil {
+				fatal(logger, fmt.Errorf("failed to close kafka consumer: %w", err))
 			}
 		}()
 
@@ -113,15 +106,13 @@ func main() {
 			topics = append(topics, topic)
 		}
 
-		consumerGroupHandler, err := adapter.NewConsumerGroupHandler(router)
-		if err != nil {
-			fatal(logger, fmt.Errorf("failed to init kafka consumer group handler: %w", err))
+		if err := c.SubscribeTopics(topics, nil); err != nil {
+			fatal(logger, fmt.Errorf("failed to subscribe to topics: %w", err))
 		}
 
 		kafkaListener, err = adapter.NewListener(
-			topics,
-			consumerGroup,
-			consumerGroupHandler,
+			c,
+			router,
 			adapter.ListenerErrorHandler(
 				transport.NewLogErrorHandler(
 					level.Error(
@@ -140,7 +131,7 @@ func main() {
 	var httpServer *http.Server
 	{
 		httpServer = &http.Server{
-			Addr:    ":8080",
+			Addr:    ":8081",
 			Handler: consumer.NewHTTPHandler(endpoints),
 		}
