+++ _examples/confluent/cmd/producer/main.go
