+++ _examples/confluent/compose.yml
@@ -27,31 +27,29 @@ services:
       - 9094:9094
 
   producer:
-    image: library/golang:1.16-alpine
+    image: library/golang:1.16
     restart: on-failure
     depends_on:
       - kafka
     environment:
       BROKER_ADDR: kafka:9092
     volumes:
-      - .:/app
-      - $GOPATH/pkg/mod:/go/pkg/mod
+      - ../..:/app
     working_dir: /app
-    command: sh -c "go mod tidy && go run ./cmd/producer"
+    command: sh -c "cd ./_examples/confluent && go mod download && go run ./cmd/producer"
     ports:
       - 8080:8080
 
   consumer:
-    image: library/golang:1.16-alpine
+    image: library/golang:1.16
     restart: on-failure
     depends_on:
       - kafka
     environment:
       BROKER_ADDR: kafka:9092
     volumes:
-      - .:/app
-      - $GOPATH/pkg/mod:/go/pkg/mod
+      - ../..:/app
     working_dir: /app
-    command: sh -c "go mod tidy && go run ./cmd/consumer"
+    command: sh -c "cd ./_examples/confluent && go mod download && go run ./cmd/consumer"
     ports:
       - 8081:8081
