+++ _examples/confluent/pkg/consumer/kafka/adapter/message.go
