+++ _examples/confluent/pkg/producer/kafka/adapter/message.go
