+++ _examples/confluent/pkg/producer/kafka/adapter/producer.go
