+++ _examples/sarama/cmd/consumer/main.go
@@ -8,8 +8,8 @@ import (
 	"os/signal"
 	"syscall"
 
+	"github.com/Shopify/sarama"
 	kafkatransport "github.com/alebabai/go-kit-kafka/kafka/transport"
-	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/go-kit/kit/transport"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
@@ -17,7 +17,7 @@ import (
 	"github.com/alebabai/go-kit-kafka/examples/common/consumer"
 	"github.com/alebabai/go-kit-kafka/examples/common/domain"
 
-	"github.com/alebabai/go-kit-kafka/examples/confluent/pkg/consumer/kafka/adapter"
+	"github.com/alebabai/go-kit-kafka/examples/sarama/pkg/consumer/kafka/adapter"
 )
 
 func fatal(logger log.Logger, err error) {
@@ -43,7 +43,6 @@ func main() {
 		logger = level.NewInjector(logger, level.InfoValue())
 		logger = log.With(logger, "ts", log.DefaultTimestampUTC)
 	}
-
 	_ = logger.Log("msg", "initialization of the application")
 
 	_ = logger.Log("msg", "initialize services")
@@ -77,23 +76,31 @@ func main() {
 
 	var kafkaListener *adapter.Listener
 	{
+		cfg := sarama.NewConfig()
+		cfg.Consumer.Offsets.Initial = sarama.OffsetOldest
+		cfg.Consumer.Offsets.AutoCommit.Enable = true
+
 		brokerAddr := domain.BrokerAddr
 		if v, ok := os.LookupEnv("BROKER_ADDR"); ok {
 			brokerAddr = v
 		}
 
-		c, err := kafka.NewConsumer(&kafka.ConfigMap{
-			"bootstrap.servers":  brokerAddr,
-			"group.id":           domain.GroupID,
-			"enable.auto.commit": true,
-		})
+		client, err := sarama.NewClient(
+			[]string{brokerAddr},
+			cfg,
+		)
 		if err != nil {
-			fatal(logger, fmt.Errorf("failed to init kafka consumer: %w", err))
+			fatal(logger, fmt.Errorf("failed to init kafka client: %w", err))
+		}
+
+		consumerGroup, err := sarama.NewConsumerGroupFromClient(domain.GroupID, client)
+		if err != nil {
+			fatal(logger, fmt.Errorf("failed to init kafka consumer group: %w", err))
 		}
 
 		defer func() {
-			if err := c.Close(); err != nil {
-				fatal(logger, fmt.Errorf("failed to close kafka consumer: %w", err))
+			if err := consumerGroup.Close(); err != nil {
+				fatal(logger, fmt.Errorf("failed to close kafka consumer group: %w", err))
 			}
 		}()
 
@@ -106,13 +113,15 @@ func main() {
 			topics = append(topics, topic)
 		}
 
-		if err := c.SubscribeTopics(topics, nil); err != nil {
-			fatal(logger, fmt.Errorf("failed to subscribe to topics: %w", err))
+		consumerGroupHandler, err := adapter.NewConsumerGroupHandler(router)
+		if err != nil {
+			fatal(logger, fmt.Errorf("failed to init kafka consumer group handler: %w", err))
 		}
 
 		kafkaListener, err = adapter.NewListener(
-			c,
-			router,
+			topics,
+			consumerGroup,
+			consumerGroupHandler,
 			adapter.ListenerErrorHandler(
 				transport.NewLogErrorHandler(
 					level.Error(
@@ -131,7 +140,7 @@ func main() {
 	var httpServer *http.Server
 	{
 		httpServer = &http.Server{
-			Addr:    ":8080",
+			Addr:    ":8081",
 			Handler: consumer.NewHTTPHandler(endpoints),
 		}
