+++ _examples/sarama/cmd/producer/main.go
