+++ _examples/sarama/pkg/consumer/kafka/adapter/consumer_group_handler.go
