+++ _examples/sarama/pkg/consumer/kafka/adapter/listener.go
