+++ _examples/sarama/pkg/consumer/kafka/adapter/message.go
