+++ _examples/sarama/pkg/producer/kafka/adapter/message.go
