+++ _examples/sarama/pkg/producer/kafka/adapter/producer.go
