--- examples/confluent/go.mod
@@ -1,11 +1,0 @@
-module github.com/alebabai/go-kit-kafka/examples/confluent
-
-go 1.16
-
-require (
-	github.com/alebabai/go-kit-kafka v1.0.0
-	github.com/alebabai/go-kit-kafka/examples/common v0.2.4
-	github.com/confluentinc/confluent-kafka-go v1.8.2
-	github.com/go-kit/kit v0.12.0
-	github.com/go-kit/log v0.2.0
-)
